[BUG #ASSIGNED 1771]: Tree drag selection does not work http://bugzilla.qooxdoo.org/show_bug.cgi?id=1771
This issue has two parts. One is a wrong method call in the Tree selectionManager class.

The other was an error in the Abstract selection manager. Once a item was selected dragging the selection doesn't work anymore. Only on new selected items the drag selection worked. With this fix we are also fixing the issue at lists and other widgets supporting drag selection.

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk/qooxdoo@17075 61aaf1ca-a10e-0410-b9fb-ac63ec25887f