Progressive: add ability to set inter-element timeout
Normally, Progressive renders its elements using a timeout of 0, which allows
dom elements to be created before rendering the next element. There are
occasional needs for a longer timeout, such as when CodeMirror must do its
initialization before subsequent code can assume it is available. This commit
adds a property to specify the inter-element timeout. This property may be
modified on a per-element basis, e.g. in the Function renderer, as needed for
specific elements.

Derrell

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk/qooxdoo@20734 61aaf1ca-a10e-0410-b9fb-ac63ec25887f