[Bug 3502] Context menu mixin "qx.ui.table.MTableContextMenu" fails to display a context menu on various Windows browsers.
Events are dispatched in different orders on Linux and Windows. On Windows, by
the time the contextmenu event was dispatched, the information about where the
user had clicked had been discarded. This commit corrects that. The cell upon
which a mousedown event occurs is saved and no longer discarded by any of the
various mouseup handlers. I've added a big "INVARIANT" comment that will
hopefully prevent people form re-adding clearing of the object which tracks
the mousedown location.

As a nice side effect of this fix, the various Table events that are generated
to the user now properly contain the row and column in which the event
occurred. Previously, the object containing the click location was cleared
BEFORE issuing the events to the user, so the row and column members of the
event data was always null.

Derrell

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk/qooxdoo@22236 61aaf1ca-a10e-0410-b9fb-ac63ec25887f