[Bug 3731] More specific exception for when same-origin test fails
This is a tricky problem, since there doesn't appear to be any way to trap
it. There's a 405 error code that Firebug sees, but there's no place in the
onreadystate handler that that value is available.

This commit adds some "suggested" problems if error code 0 occurs. It tests to
see if the current page was loaded with a file: URL (a common cause of
questions on the mailing list), and if not, then it suggests that possibly
a cross-domain request was issued. This should at least get people thinking
along the right direction to more quickly solve the problem even though it's
not fool-proof.

Derrell

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk/qooxdoo@22239 61aaf1ca-a10e-0410-b9fb-ac63ec25887f