[Bug 2866] Inline edit in VirtualTree is not working
Step 1 to fix this bug:

- Remove the special handling of handleMoveKeyDown.

- In the _handleSelectEvent() method, pass control to the superclass
  immediately if the focused cell is not in the tree column.

- While I'm at it, update from 0.6 methodology of calling the superclass.

This is an intermediate step in the complete correction of the problem, as I
currently have enabled the focus indicator. It must be on for editing (true in
Table as well). The next step will be to make the focus indicator invisible
when it is not desired to be displayed, rather than removing it completely.

Derrell

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk/qooxdoo@22273 61aaf1ca-a10e-0410-b9fb-ac63ec25887f