Allow table model to affect the table
A table model is passed to the constructor of a table, so the table can affect
the table model, but the table model previously had no way to affect
properties of the table. This commit adds an optional init() method in the
table model, which is called after a new table model is applied. The init()
method is passed the table object, so can affect changes upon the table.

This is used, for example, by the Smart table model, to change the value of
the resetSelectionOnHeaderClick property.

Derrell

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk/qooxdoo@22501 61aaf1ca-a10e-0410-b9fb-ac63ec25887f