[Bug 3803] qx.ui.table.pane.Pane._getRowsHtml cellInfo.style is not reseted between calls
Although cellInfo.style is documented as part of the interface for a cell
renderer, it was not set anyplace. There is now a new property in
Table.cellrenderer.Abstract called defaultCellStyle that defaults to null for
backward compatibility. Users may set this property to provide a style to be
used by the cell renderer each time the cell is rendered. The code in Pane
which calls the cell renderer now resets cellInfo.style to the (current!)
defaultCellStyle value on each iteration of the loop.

Derrell

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk/qooxdoo@22529 61aaf1ca-a10e-0410-b9fb-ac63ec25887f