[Bug 4568] table-contextmenu displayed on gray area of the table
The availability of the context menu in the area of the table below the (small
set of) data rows was inconsistent and depended on the setting of the property
showCellFocusIndicator. The behavior has now been made consistent, and a new
property, contextMenuFromDataCellsOnly controls whether the cellContextmenu
event is fired upon right-click in the area beneath the data rows.

Derrell

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk/qooxdoo@24667 61aaf1ca-a10e-0410-b9fb-ac63ec25887f