[BUG #4607] Auto-magic or explicit parsing of response
By default, the response is parsed according to the content type of the response (if a parser can be found). The raw response is still accessible with getResponseText().

As an alternative, a parser can be forced - either by passing a symbolic string that represents a known parser or a function that gets passed the raw response when parsing.

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk/qooxdoo@26845 61aaf1ca-a10e-0410-b9fb-ac63ec25887f