[BUG #5104] Support request.Xhr alongside remote.Request
Store can be configured to use request.Xhr with setDeprecatedTransport(false). By default, the store uses the old transport. When the new transport is used the API changes in some subtle details. Therefore, it is best to make the user change the transport knowingly. Deprecation warnings are thrown if the old transport is used.

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk/qooxdoo@27283 61aaf1ca-a10e-0410-b9fb-ac63ec25887f