[Bug 5109] instanceof Array sometimes doesn't detect an array (rhino compiled)
instanceof Array (or anything else, for that matter) requires a common
prototype to return true. As explained in
http://perfectionkills.com/instanceof-considered-harmful-or-how-to-write-a-robust-isarray/
one can't always count on having a common prototype. Two known examples of
such are multi-frame DOM environments, and rhino-compiled classes.

This commit replaces occurrances of
  x instanceof Array
with
  qx.Bootstrap.getClass(x) === "Array"

which uses a reliable method of determining that the object is in fact an
Array.

Derrell

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk/qooxdoo@27427 61aaf1ca-a10e-0410-b9fb-ac63ec25887f